feat: migrated from JSON to SQLite

+++ setup.sh
@@ -1,15 +1,15 @@
+#!/usr/bin/env bash
+set -e
+
 python3 -m venv venv
 source venv/bin/activate
 
 pip install --upgrade pip
 pip install -r requirements.txt
 
-cat > .env << EOF
+if [ ! -f .env ]; then
+    cat > .env << EOF
 BOT_TOKEN=""
-PASSWORD=""
 API_KEY=""
 EOF
-
-mkdir resources/
-touch resources/users.json
-touch resources/plates.json
+fi
